requirements.txt
appdirs==1.4.4
astroid==2.12.13
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.8.0
Flask==2.2.2
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
huggingface-hub==0.11.1
idna==3.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==9.0.0
mypy-extensions==0.4.3
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
pathspec==0.10.2
platformdirs==2.5.4
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.15.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.0
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.16.0
stable-ts==1.0.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
whisper==1.0
wrapt==1.14.1
